+++ src/enclave/Enclave/CMakeLists.txt
@@ -10,6 +10,7 @@ set(SOURCES
   Flatbuffers.cpp
   FlatbuffersReaders.cpp
   FlatbuffersWriters.cpp
+  IntegrityUtils.cpp
   Join.cpp
   Project.cpp
   Sort.cpp
