+++ src/enclave/Enclave/FlatbuffersReaders.h
@@ -1,13 +1,11 @@
 #include "Flatbuffers.h"
-#include "EnclaveContext.h"
+#include "IntegrityUtils.h"
 
 #ifndef FLATBUFFERS_READERS_H
 #define FLATBUFFERS_READERS_H
 
 using namespace edu::berkeley::cs::rise::opaque;
 
-void init_log(const tuix::EncryptedBlocks *encrypted_blocks);
-void verify_log(const tuix::EncryptedBlocks *encrypted_blocks, std::vector<LogEntry> past_log_entries);
 /**
  * A reader for Row objects within an EncryptedBlock object that provides both iterator-based and
  * range-style interfaces.
